LayoutTests/imported/w3c/web-platform-tests/css/selectors/invalidation/is-expected.txt

Red
Red
Green
Green
Blue
Blue
Red
Black
Red

PASS Preconditions. 
FAIL Invalidate :is() for simple selector arguments. assert_equals: expected "rgb(255, 0, 0)" but got "rgb(255, 255, 0)"
FAIL Invalidate :is() for compound selector arguments. assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
FAIL Invalidate :is() for complex selector arguments. assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
FAIL Invalidate nested :is(). assert_equals: expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
FAIL Test specificity of :is(). assert_equals: expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
